+++ ReLiveWP.Services.Push/Data/NotificationStatus.cs
@@ -1,0 +1,10 @@
+namespace ReLiveWP.Services.Push.Data;
+
+public enum NotificationStatus
+{
+    Pending,
+    InFlight,
+    Delivered,
+    Expired,
+    Zombied,
+}
