+++ ReLiveWP.Services.Push/Session/PushSessionContext.cs
@@ -1,0 +1,6 @@
+namespace ReLiveWP.Services.Push.Session;
+
+public class PushSessionContext
+{
+    public string DeviceId { get; set; }
+}
