+++ ReLiveWP.Services.Push/Session/PushSessionStateType.cs
@@ -1,0 +1,9 @@
+namespace ReLiveWP.Services.Push.Session;
+
+public enum PushSessionStateType
+{
+    Connecting,
+    Reconnecting,
+    Connected,
+    Errored
+}
